requirements.txt
numpy
opencv-python
addict
six
torch>=1.1.0
torchvision>=0.3.0
pillow<7.0
lmdb
nltk
terminaltables
albumentations
